Add fork as false so that we will be able to debug
Besides, it seems that devTools will be disabled (because of fork=false)
but it's updating the classpath with every change so that's weird.
It's also ignoring the arguments, so it seems that it's not required
for debugging (so neither the remote config).
Refer to this doc for more info: https://access.redhat.com/documentation/en-us/red_hat_support_for_spring_boot/2.1/html/spring_boot_2.1.x_runtime_guide/debugging-springboot-based-application_spring-boot

PS: Bear in mind that we should always use debug mode for all of this to work.